Building with the html tags will now also load templates from disk
Templates were previously always only parsed and loaded upon start,
this made any form of template tweaking slow. Now it will parse them
on every request if the binary was build using the tag "html".
